+++ fsl/utils/tempdir.py
@@ -5,12 +5,13 @@
 # Author: Paul McCarthy <pauldmccarthy@gmail.com>
 #
 """This module contains utilities for working with temporary files and
-directories. It currently only contains one function:
+directories. It currently contains the following functions:
 
 .. autosummary::
    :nosignatures:
 
    tempdir
+   mkstemp
 """
 
 
@@ -21,7 +22,7 @@ import contextlib
 
 
 @contextlib.contextmanager
-def tempdir(root=None, changeto=True, override=None, prefix=None):
+def tempdir(root=None, changeto=True, override=None, prefix=None, delete=True):
     """Returns a context manager which creates and returns a temporary
     directory, and then deletes it on exit.
 
@@ -35,10 +36,15 @@ def tempdir(root=None, changeto=True, override=None, prefix=None):
 
     :arg override: Don't create a temporary directory, but use this one
                    instead. This allows ``tempdir`` to be used as a context
-                   manager when a temporary directory already exists.
+                   manager when a temporary directory already exists. Implies
+                   ``delete=False``.
 
     :arg prefix:   Create the temporary directory with a name starting with
                    this prefix.
+
+    :arg delete:   If ``True`` (the default), the directory is deleted on exit.
+                   Otherwise the caller is responsible for deleting the
+                   directory.
     """
 
     if root is not None:
@@ -48,6 +54,7 @@ def tempdir(root=None, changeto=True, override=None, prefix=None):
         testdir = tempfile.mkdtemp(dir=root, prefix=prefix)
     else:
         testdir = override
+        delete  = False
 
     prevdir = os.getcwd()
 
@@ -57,7 +64,16 @@ def tempdir(root=None, changeto=True, override=None, prefix=None):
         yield testdir
 
     finally:
-        if override is None:
+        if delete:
             shutil.rmtree(testdir)
         if changeto:
             os.chdir(prevdir)
+
+
+def mkstemp(*args, **kwargs):
+    """Wrapper around ``tempfile.mkstemp``. Does the same as that
+    function, but closes the file, and just returns the file name.
+    """
+    hd, fname = tempfile.mkstemp(*args, **kwargs)
+    os.close(hd)
+    return fname